Move cache
Prompt: "Move cache.py and cache_test.py to the utilities directory, renaming cache_test.py to test_cache.py."
